io read/write locking to protect [SRAM Start Address] and [SRAM data] registers.
This registers are shared by all 8 channels of the submodule.
In order to safely read/write function vectors in each of the
channels -- locking should be used.

Otherwice we can fall in the sutiation, when one user set this 2 registers
to read/write data in one of the channels for specified function can
be postponed by another user, which in turn will set up 2 registers
to read/write from another channel.
When first user will be woken up to continue his r/w operation -- register
setting will be spoiled by the second user.

Signed-off-by: Yury Georgievskiy <ygeorgie@cern.ch>