vmodio: fix isr bug
This patch addresses a serious bug in IRQ handler register. The irq
number is passed as the void * argument to the handler so that it
recognizes which card and slot caused the interrupt. But, in calling
vme_request_irq, the address of a local, automatic variable was
provided.

Now, the four IRQ numbers for a card are stored in its struct field
irq, so that a correct address with lifetime longer than automatic
is passed to the handler