cvorg/driver: establish a limitation of 100 MHz of the sampling clock frequency.
CVORG has been designed to use a sampling clock =< 100 MHz.

The problem appeared originally when the user have an external clock of 100.002 MHz, so
in the formula, this value is rounded up to 101 MHz and the minimum number of samples
goes to 23895 (above the limit of 23200).

After a meeting with the HW engineer, we decided that it is better to limit by SW
that the maximum allowed frequency is 100 MHz to avoid unexpected behaviors in
the board.

We detected that there is an error of 2 kHz in the measurement of the external
clock at 100 MHz.

The other cases, we have errors about: 1.4 kHz at 75 MHz, 0.9 kHz at 50 MHz,
0.5 kHz at 25 MHz.

But the extrapolation of the minimum number of samples goes to safer values
with these errors, so it is not critical. However, in the maximum frequency we
need to take in account the measurement error.

Signed-off-by: Samuel Iglesias Gonsalvez <samuel.iglesias.gonsalvez@cern.ch>