cvorb: fix bug in multiple LUN access
In setting up the user statics structure of a CVORB module,
the pointer referring to it was initialized to a statically allocated
constant address. The net result is that each module installation
overwrites the user statics structure of the former; every subsequent
access to a module will ignore the LUN used as a parameter and
go directly to the module last installed. This fixes the problem
(hopefully) by dynamically allocating the structure