cvorb: delete static struct pll and use a normal access
The bug is present when the user has several CVORB boards installed on the box.
After the initialization of the driver, if the user wants to setup the DAC or
the PLL, he accesses to the same board always.

The reason is the static struct pll used in ad9516o.c that overwrites the previous
setup. The rest of registers are well defined and used, so the bug was not easy to spot.

Now, the base address of the desired board is passed as a parameter, hidden in a
CvorbUserStatics_t type of variable. The base address is the same that the first
register on the bank 0 ("Because I'm worth it.").

Waiting for a new driver, this patch will do the job.

	"Terrific problems require horrific solutions."
		S.I.G., Nov 2011

Signed-off-by: Samuel Iglesias Gonsalvez <siglesia@cern.ch>
Signed-off-by: Juan David Gonzalez Cobas <dcobas@cern.ch>