cvorg-new: added new driver for CVORG board
This driver is created from scratch being "kosher", i.e, removing all the skel
dependencies.

Also the driver is designed to facilitate a possible port to ZIO framework.

It includes a new install script, new library (maintaining the same API
but removing some functionality like the cvorg_lock, cvorg_unlock functions)
and a rewriting of the extest framework to use the same test program used
before.

The extest framework should be rewritten to be generic and not only for
the cvorg. This is a work to be done.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesia@cern.ch>