vmod/vmodttl: changed the support of IRQ for one linked list to only the last one.
This board has problems with a huge number of IRQ. If the IRQ arising frequency
is higher of 1 kHz, the board will give IRQ in channels that it is not
configured (!?)

Also, the user wanted that the applications waiting for an IRQ received the
same IRQ all of them when there is one, and return all of them at the same
time. This is the reason of the extra waiting queue in the vmodttl_read
function.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesia@cern.ch>