ctr: cdcm/semaphore: remove kthread_should_stop() call
The call in kthread_should_stop() was added to help in the
emulation of Lynx' kthreads; however I doubt that the emulation
was 1) ever tested or 2) ever had any users.

This call seemed to work fine on 2.6.24; however Julian
has detected an oops when calling it from the ctr driver
on 2.6.33-rt:

Call Trace:
 [<ffffffff8135b658>] panic+0x89/0x143
 [<ffffffff8135eb38>] oops_end+0xae/0xbe
 [<ffffffff81025a22>] no_context+0x1fc/0x20b
 [<ffffffff81025baf>] __bad_area_nosemaphore+0x17e/0x1a1
 [<ffffffff81025be5>] bad_area_nosemaphore+0x13/0x15
 [<ffffffff81360651>] do_page_fault+0x140/0x299
 [<ffffffff8135e04f>] page_fault+0x1f/0x30
 [<ffffffff8105d489>] ? kthread_should_stop+0x19/0x1e
 [<ffffffff8135d865>] ? rt_spin_lock+0xe/0x10
 [<ffffffffa02d04c8>] __cdcm_down_common+0xb2/0xf5 [ctrv]
 [<ffffffffa02d055f>] __cdcm_down_interruptible+0x1d/0x1f [ctrv]
 [<ffffffffa02d05e8>] swait+0x87/0xa1 [ctrv]
 [<ffffffffa02d3b0f>] CtrDrvrRead+0x146/0x25d [ctrv]
 [<ffffffffa02d44f3>] LynxOsRead+0x92/0x123 [ctrv]
 [<ffffffff810faaa3>] ? rw_verify_area+0x8d/0xb0
 [<ffffffff810fb017>] vfs_read+0xab/0x108

Since Lynx' kthread emulation is probably broken from day 1,
the appended removes the call to kthread_should_stop().

NB: This patch is UNTESTED. IIRC the call to kthread_should_stop()
was the only modification we did over the original implementation
in the kernel, so this "should" work.

Signed-off-by: "Emilio G. Cota" <cota@braap.org>
Reported-by: "Julian Lewis" <lewis@cern.ch>