ctr: set up new ioctl() numbers
The original second arguments of ioctl() in the ctr driver
were directly taken from the LynxOS implementation, and were
sequential numbers from 1 on. This breaks completely
after 2.6.29, so a fix is implemented: ioctl() in this driver
will accept both the old numbers and the canonical, new numbers
constructed via the _IO[RW] macros.

This patch declares the two sets of ioctl numbers