ctr: change everything so that it all stays the same
This commit registers the deployment of the ctr driver
done in December 2011, where all the Linux versions
of the ctr driver in lab, oper and oplhc were levelled
and track to this original sources.

This version of the driver is the starting point to
begin tracking correctly all its modifications.

With this, the correct driver deployed in all the lab
and operation machines is correctly tracked for good.