ipoctal: fix tx logic
This fixes two bogus bits. First, nr_bytes is unsigned, so the
check nr_bytes <= 0 only makes sense to verify that it is actually
zero. Second, when this happens, the conditional precludes waking
up the client that asked for the transmission, if any. The right
thing to do is simply to bypass the writing of a character if nr_bytes
is 0.