vd80: allow /dev/vd80.0 to exist
The library routine vd80OpenHandle tries to open any of
16 device nodes /dev/vd80.n, n = 1..16. There is no
reason why /dev/vd80.0 could not be treated equally.
Skel treats the module number as a numerical tag
with no significance as index in any table, and many
skel drivers conform to the convention that a module
number can be zero.

So lun=0 is a valid insmod argument, and vd80.awk need not
increment the LUN number artificially.