ctr: CtrReadInfo: bugfix for multiple modules
When multiple modules are configured in a readinfo
file, the current row number for the involved module
needs updated. The module number was not being updated
correctly when several modules were described in a readinfo
file, and this fixes the issue. This applies especially
when several readinfo files are applied in a sequence,
as their contents are appended to the action tables
of the modules.

Signed-off-by: Julian Lewis <julian.lewis@cern.ch>