ipoctal: ipoctal_shutdown callback moved to tty_port operations
Before this change some things seemed to remain open after a
communication close - shutdown call - : the second time we tried to
open a device we get an I/O error.

If a custom shutdown callback is present at tty level, we
are responsible to take care of removing tty and releasing termios.
Nothing of that was present in current code, so I *guess* that
implemented shutdown operations should be done at tty_port level, not
tty. Then I move the callback from tty ops to tty_port ops. It is
working fine now but I am not still sure if this is conceptually right.

Also ipoctal_shutdown is renamed to ipoctal_port_shutdown