cvora: add new modes
An entire set of new modes of operation has been added by
the hardware designer, among them the "up only" mode requested
by JMN, and CVORB protocol modes.

This adds the constants to the library and to the test
program. In addition, the mode bit mask is enlarged to cater
for up to 256 modes, with the foresight of a potential
(continuum?) infinity of modes to be supported